--- a/lib/compiler/resinator/main.zig
+++ b/lib/compiler/resinator/main.zig
@@ -50,12 +50,6 @@ pub fn main() !void {
         },
     };
 
-    if (zig_integration) {
-        // Send progress with a special string to indicate that the building of the
-        // resinator binary is finished and we've moved on to actually compiling the .rc file
-        try error_handler.server.serveStringMessage(.progress, "<resinator>");
-    }
-
     var options = options: {
         var cli_diagnostics = cli.Diagnostics.init(allocator);
         defer cli_diagnostics.deinit();