… lib/std/debug.zig
… lib/std/fmt.zig
… lib/std/io/tty.zig
… lib/std/json/dynamic.zig
… lib/std/log.zig
… lib/std/os/windows/kernel32.zig
… lib/std/process.zig
… lib/std/process/Child.zig
… lib/std/zig.zig
… lib/std/zig/ErrorBundle.zig
… lib/std/zig/Server.zig
… src/Compilation.zig
… src/Module.zig
… src/Package/Fetch.zig
… src/glibc.zig
… src/libcxx.zig
… src/libtsan.zig
… src/libunwind.zig
… src/link.zig
… src/link/C.zig
… src/link/Coff.zig
… src/link/Coff/lld.zig
… src/link/Elf.zig
… src/link/MachO.zig
… src/link/NvPtx.zig
… src/link/Plan9.zig
… src/link/SpirV.zig
… src/link/SpirV/deduplicate.zig
… src/link/SpirV/lower_invocation_globals.zig
… src/link/SpirV/prune_unused.zig