## Demo

[simple asciinema demo](https://asciinema.org/a/gDna9RnicwYjDRIDn4e07NFSc) [source](https://gist.github.com/andrewrk/b22b4f663cef6b4672d7097de95ea343)

[demo: building a music player with zig build](https://asciinema.org/a/Hld34XJAS2enQZVCsh7WeVAlq) [source](https://codeberg.org/andrewrk/player)

## Motivation

The previous implementation of `std.Progress` had the design limitation that it could not assume ownership of the terminal. This meant that it had to play nicely with sub-processes purely via what was printed to the terminal, and it had to play nicely with progress-unaware stderr writes to the terminal. It also was forbidden from installing a SIGWINCH handler, or running ioctl to find out the rows and cols of the terminal.

This implementation is designed around the idea that a single process will be the sole owner of the terminal, and all other progress reports will be communicated back to that process. With this change in the requirements, it becomes possible to make a much more useful progress bar.

## Strategy

This creates a standard "Zig Progress Protocol" and uses it so that the same `std.Progress` API works both when an application is the main owner of a terminal, and when an application is a child process. In the latter case, progress information is communicated semantically over a pipe to the parent process.

The file descriptor is given in the `ZIG_PROGRESS` environment variable. `std.process.Child` integrates with this, so attaching a child's progress subtree in a parent process is as easy as setting the `child.progress_node` field before calling `spawn`.

In order to avoid performance penalty for using this API, the `Node.start` and `Node.end` APIs are thread-safe, lock-free, infallible, and do minimal amount of memory loads and stores. In order to accomplish this, a statically allocated buffer of `Node` storage is used - one array for parents, and one array for the rest of the data. Children are not stored. The statically allocated buffer is used for a bespoke `Node` allocator implementation. A static buffer is sufficient because we can set an upper bound on supported terminal width and height. If the terminal size exceeds this, the progress bar output will be truncated regardless.

A separate thread periodically refreshes the terminal on a timer. This progress update thread iterates over the entire preallocated parents array, looking for used nodes. This is efficient because the parents array is only 200 8-bit integers, or about 4 cache lines. When iterating, this thread "serializes" the data into a separate preallocated array by atomically loading from the shared data into data that is only touched by a single thread - the progress update thread. It then looks for nodes that are marked with a file descriptor that is a pipe to a child process. Such nodes are replaced during the serialization process with the data from reading from the pipe. The data can be memcpy'd into place except for the parents array which needs to be relocated. Once this serialization process is complete, there are two paths, one for a child process, and one for the root process that owns the terminal.

The root process that owns the terminal scans the serialized data, computing children and sibling pointers. The canonical data only stores parents, so this is where the tree structure is computed. Then the tree is walked, appending to a static buffer that will be sent to the terminal with a single write() syscall. During this process, the detected rows and cols of the terminal are respected. If the user resizes the terminal, it will cause a SIGWINCH which signals the update thread to wake up and redraw with the new rows and cols.

A child process, instead of drawing to the terminal, takes the same serialized data and sends it across a pipe. The pipe is in non-blocking mode, so if it fills up, the child drops the message; a future update will contain the new progress information. Likewise when the parent reads from the pipe, it discards all messages in the buffer except for the last one. If there are no messages in the pipe, the parent uses the data from the last update.

## Performance Data Points

Building the self-hosted compiler as a sub-process. This measures the cost of the new API implementations, primarily `Node.start`, `Node.end`, and the disappearance of `Node.activate`. In this case, standard error is not a terminal, and thus an update thread is never spawned.

```
Benchmark 1 (3 runs): master branch zig build-exe self-hosted compiler
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          51.8s  ±  138ms    51.6s  … 51.9s           0 ( 0%)        0%
  peak_rss           4.57GB ±  286KB    4.57GB … 4.57GB          0 ( 0%)        0%
  cpu_cycles          273G  ±  360M      273G  …  274G           0 ( 0%)        0%
  instructions        487G  ±  105M      487G  …  487G           0 ( 0%)        0%
  cache_references   19.0G  ± 31.8M     19.0G  … 19.1G           0 ( 0%)        0%
  cache_misses       3.87G  ± 12.0M     3.86G  … 3.88G           0 ( 0%)        0%
  branch_misses      2.22G  ± 3.44M     2.21G  … 2.22G           0 ( 0%)        0%
Benchmark 2 (3 runs): this branch zig build-exe self-hosted compiler
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          51.5s  ±  115ms    51.4s  … 51.7s           0 ( 0%)          -  0.4% ±  0.6%
  peak_rss           4.58GB ±  190KB    4.58GB … 4.58GB          0 ( 0%)          +  0.1% ±  0.0%
  cpu_cycles          272G  ±  494M      272G  …  273G           0 ( 0%)          -  0.4% ±  0.4%
  instructions        487G  ± 63.5M      487G  …  487G           0 ( 0%)          -  0.1% ±  0.0%
  cache_references   19.1G  ± 16.9M     19.1G  … 19.1G           0 ( 0%)          +  0.3% ±  0.3%
  cache_misses       3.86G  ± 17.1M     3.84G  … 3.88G           0 ( 0%)          -  0.2% ±  0.9%
  branch_misses      2.23G  ± 5.82M     2.22G  … 2.23G           0 ( 0%)          +  0.4% ±  0.5%
```

Building the self-hosted compiler, with `time zig build-exe -fno-emit-bin ...` so that the progress is updating the terminal on a regular interval.

* Old:
  * 0m4.115s
  * 0m4.216s
  * 0m4.221s
  * 0m4.227s
  * 0m4.234s
* New (1% slower):
  * 0m4.231s
  * 0m4.240s
  * 0m4.271s
  * 0m4.339s
  * 0m4.340s

Building my music player application with `zig build`, with the project-local cache cleared. This displays a lot of progress information to the terminal. This data point accounts for many sub processes sending progress information over a pipe to the parent process for aggregation.

* Old
  * 65.74s
  * 66.39s
  * 71.09s
* New (1% faster)
  * 65.51s
  * 65.88s
  * 66.09s

## Upgrade Guide

* Pass around `std.Progress.Node` instead of `*std.Progress.Node` (no longer a pointer).
* Remove calls to `node.activate()`. Those are not needed anymore.
* It's now safe to pass short-lived strings to `Node.start` since the data will be copied.
* It is illegal to initialize `std.Progress` more than once. Do that in main() and nowhere else.
* Use `std.debug.lockStdErr` and `std.debug.unlockStdErr` before writing to stderr to integrate properly with `std.Progress` (`std.debug.print` already does this for you).

```diff
-    var progress = std.Progress{
-        ...
-    };
-    const root_node = progress.start("Test", test_fn_list.len);
+    const root_node = std.Progress.start(.{
+        .root_name = "Test",
+        .estimated_total_items = test_fn_list.len,
+    });
```

All the options to `start` are optional.

Finally, when spawning a child process, populate the `progress_node` field first:

```zig
child.progress_node = node;
try child.spawn();
```

## Merge Checklist

* follow-up work
  * implement single-threaded mode
  * add the concept of byte unit flag and resource scope; support throughput calculation
  * on WASI use `/ZIG_PROGRESS` preopen instead of env var
  * show (+ X more nodes) when truncating based on rows?
  * Windows IPC support

closes #9633
closes #17821