@VisenDev 2024-05-25
Very nice! This would probably close this issue
https://github.com/ziglang/zig/issues/17821