@andrewrk 2024-05-25
@squeek502 do you have any interest in helping out on the Windows side of things here? There are two challenges:

1. The logic in `computeRedraw` works when the terminal supports ANSI escape codes, but it does not work for the case when we need to use kernel32 console function calls. The task here is to add a Windows-specific implementation for rendering this tree data structure to the terminal, as well as erasing the previous render.
2. Child process integration. This implementation sets up a child with file descriptor `3` as a pipe for communicating progress to the parent on POSIX-like systems. On Windows we need to do some kind of equivalent.

I think the second one is the more difficult one, but even with only (1) solved, it would still be useful. There would not be visibility into child processes but it would be a good start.