@squeek502 2024-05-26
> it might be a few days before I can give it proper attention

I lied; I worked on the Windows console API stuff and got a proof-of-concept working but it's a bit weird. The implementation is here:

https://github.com/squeek502/zig/commit/78a10937192376647b126216a41868f796823ebb

(all the IPC stuff is commented out as a quick fix to get it to compile on Windows)

When the console supports ANSI escape codes, then ANSI escape codes are used and it works as expected:

https://github.com/ziglang/zig/assets/2389051/7edd3029-ed2d-4759-80e1-245331d2b9bd

But when the console does not support ANSI escape codes, it falls back to the Windows console API:

https://github.com/ziglang/zig/assets/2389051/a65b0995-fd7e-4a1e-b338-1df51b705948

(the flickering only happened while I was recording with OBS; there's very minimal flickering when not recording)

Note that there is a ► character at the start of the progress when using the Windows API. This is because AFAICT there's no way to go to/find the start of a line if the line is being wrapped. That is, carriage returns only bring you to the start of the row, not the start of the line. So, the workaround is to write out a special character at the start of the progress and then take advantage of the `ReadConsoleOutputCharacter` API to find the marker and clear the screen starting from there. This makes it so that resizing/wrapping doesn't break the rendering.

The important part of the implementation is here:

https://github.com/squeek502/zig/blob/78a10937192376647b126216a41868f796823ebb/lib/std/Progress.zig#L524-L577

with some commented out alternate implementations below it that worked somewhat but had issues.

Also, when falling back to the Windows console API it currently assumes [code page 437](https://en.wikipedia.org/wiki/Code_page_437) (the default code page for consoles) is the active code page and uses that for the box drawing characters. Some TODOs about that are here:

https://github.com/squeek502/zig/blob/78a10937192376647b126216a41868f796823ebb/lib/std/Progress.zig#L471-L481