@squeek502 2024-05-28
Fixed up and rebased my Windows console API stuff on top of the latest commits here: https://github.com/squeek502/zig/commit/5b49900438ccedb5cae6031b5979822a32cee507

(note: it's a different branch than the one in my previous comment)

Feel free to cherry pick the commit, or use it as reference and adapt it however you'd like.