@andrewrk 2024-05-28
> Hmm, the windows implementation has the undesirable property that it clears the screen, including all scrollback. Ideally it would not do that. In fact I think it's better to show no progress than to destroy scrollback history.

Might be an artifact from using wine. Haven't tried on my Windows laptop yet