@squeek502 2024-05-28
> Hmm, the windows implementation has the undesirable property that it clears the screen, including all scrollback

Can't reproduce this behavior on Windows.

In testing this, though, I have noticed that what I said in https://github.com/ziglang/zig/pull/15206#issuecomment-1556633886 / https://github.com/ziglang/zig/pull/18692#issuecomment-1912516673 appears to be wrong. With a `togglevt.zig` like this:

```
const std = @import("std");
const windows = std.os.windows;

pub fn main() void {
    const stderr = std.io.getStdErr();
    var mode: windows.DWORD = undefined;
    std.debug.assert(std.os.windows.kernel32.GetConsoleMode(stderr.handle, &mode) != 0);

    const new_mode = mode ^ windows.ENABLE_VIRTUAL_TERMINAL_PROCESSING;
    std.debug.assert(std.os.windows.kernel32.SetConsoleMode(stderr.handle, new_mode) != 0);

    const now_enabled = new_mode & windows.ENABLE_VIRTUAL_TERMINAL_PROCESSING != 0;
    std.debug.print("{s} virtual terminal processing for stderr: {}\n", .{ if (now_enabled) "enabled" else "disabled", stderr.handle });
}
```

then running it in Windows Terminal results in:

```
C:\Users\Ryan\Programming\Zig\tmp\progress> togglevt
disabled virtual terminal processing for stderr: anyopaque@290

C:\Users\Ryan\Programming\Zig\tmp\progress> togglevt
disabled virtual terminal processing for stderr: anyopaque@290

C:\Users\Ryan\Programming\Zig\tmp\progress> togglevt
disabled virtual terminal processing for stderr: anyopaque@290
```

(that is, the terminal is resetting the console mode to the default for each run; the mode is not preserved between runs which was what I thought happened)

With `cmd.exe`, the same thing happens except `ENABLE_VIRTUAL_TERMINAL_PROCESSING` is disabled by default:

```
C:\Users\Ryan\Programming\Zig\tmp\progress>togglevt
enabled virtual terminal processing for stderr: anyopaque@58

C:\Users\Ryan\Programming\Zig\tmp\progress>togglevt
enabled virtual terminal processing for stderr: anyopaque@58
```

So another option here might be to revisit those PRs (I like https://github.com/ziglang/zig/pull/18692 best) and try setting `ENABLE_VIRTUAL_TERMINAL_PROCESSING` within `File.supportsAnsiEscapeCodes`. If we go that route, we might be able to drop Windows console API support entirely since [all the console API docs have a "use ANSI escape codes instead" notice on them](https://learn.microsoft.com/en-us/windows/console/fillconsoleoutputcharacter) and `ENABLE_VIRTUAL_TERMINAL_PROCESSING` was added in Windows 10 RS1 from June 2016.